requirements.txt
pygame==2.5.2
websocket-client==1.7.0